backend/deal_users/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
boto3==1.26.135
botocore==1.29.135
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
dataclasses-json==0.5.7
Django==4.2.4
django-cors-headers==3.14.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-ratelimit==4.0.0
django-storages==1.13.2
djangorestframework==3.14.0
faiss-cpu==1.7.3
frozenlist==1.3.3
google-search-results==2.4.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
langchain==0.0.136
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pdfminer.six==20221105
phonenumbers==8.13.10
Pillow==9.5.0
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==1.4.47
sqlparse==0.4.3
tenacity==8.2.2
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
uWSGI==2.0.21
yarl==1.8.2
djangorestframework-api-key==2.3.0
matplotlib==3.7.1
daphne==4.0.0
Twisted==22.10.0
channels==4.0.0
pycryptodome==3.18.0
Twisted[http2,tls]
reportlab==4.0.4
stripe==6.2.0
oauthlib==3.2.2
google-analytics-data==0.17.2
celery==5.3.5
geoip2==4.7.0
redis==4.5.5
django-redis==5.2.0
google-auth==2.23.3